Update kafka consumer to support IAM auth
