+++ server/kafka/consumer.go
@@ -95,6 +95,10 @@ func NewConsumer(cc conf.ConnectorConfig, dialTimeout time.Duration) (Consumer, 
 		}
 		sc.Net.SASL.User = cc.SASL.User
 		sc.Net.SASL.Password = cc.SASL.Password
+	} else if cc.IAM.Enable && cc.IAM.Region != "" {
+		sc.Net.SASL.Enable = true
+		sc.Net.SASL.Mechanism = sarama.SASLTypeOAuth
+		sc.Net.SASL.TokenProvider = &MSKAccessTokenProvider{Region: cc.IAM.Region}
 	}
 
 	if sc.Net.SASL.Enable && cc.SASL.InsecureSkipVerify {
@@ -105,6 +109,10 @@ func NewConsumer(cc conf.ConnectorConfig, dialTimeout time.Duration) (Consumer, 
 	} else if tlsC, err := cc.TLS.MakeTLSConfig(); tlsC != nil && err == nil {
 		sc.Net.TLS.Enable = true
 		sc.Net.TLS.Config = tlsC
+	} else if cc.IAM.Enable {
+		tlsConfig := tls.Config{}
+		sc.Net.TLS.Enable = true
+		sc.Net.TLS.Config = &tlsConfig
 	}
 
 	if cc.MinBytes > 0 {
