+++ server/kafka/iam_token_provider.go
@@ -1,0 +1,17 @@
+package kafka
+
+import (
+	"context"
+
+	"github.com/Shopify/sarama"
+	"github.com/aws/aws-msk-iam-sasl-signer-go/signer"
+)
+
+type MSKAccessTokenProvider struct {
+	Region string // AWS IAM region
+}
+
+func (m *MSKAccessTokenProvider) Token() (*sarama.AccessToken, error) {
+	token, _, err := signer.GenerateAuthToken(context.TODO(), m.Region)
+	return &sarama.AccessToken{Token: token}, err
+}
